Changes to RumaResponder
This commit adds a new variant to the `RumaResponder` enum, `OneOff`. This variant allows for custom responses that don't fit the typical matrix error structure to be returned from endpoints with an associated status code.

For an example of such a response body, refer to the 401 response for the /v3/register endpoint in the client-server API

https://spec.matrix.org/latest/client-server-api/#post_matrixclientv3register