Created `parquet_file!` macro
This commit adds the `parquet_file!` to `cubby_server::utils`. This macro is an easy way to create lots of parquet files without much syntax clutter for use in development and testing.

This macro should most likely be phased out before this server will be in any seriously usable state.